third_party/WebKit/LayoutTests/external/wpt/fetch/api/request/request-reset-attributes.https-expected.txt

This is a testharness.js-based test.
FAIL Request.isReloadNavigation is reset with non-empty RequestInit assert_equals: expected "old: false, new: false" but got "hello\n"
FAIL Request.isHistoryNavigation is reset with non-empty RequestInit assert_equals: expected "old: false, new: false" but got "{\"error\": {\"message\": \"\", \"code\": 404}}"
FAIL Request.mode is reset with non-empty RequestInit when it's "navigate" assert_equals: expected "old: navigate, new: same-origin" but got "hello\n"
Harness: the test ran to completion.
